+++ aerich/migrate.py
@@ -13,6 +13,7 @@ from tortoise import BaseDBAsyncClient, Model, Tortoise
 from tortoise.exceptions import OperationalError
 from tortoise.indexes import Index
 
+from aerich.coder import load_index
 from aerich.ddl import BaseDDL
 from aerich.models import MAX_VERSION_LENGTH, Aerich
 from aerich.utils import (
@@ -245,6 +246,8 @@ class Migrate:
         for x in cls._handle_indexes(model, model_describe.get("indexes", [])):
             if isinstance(x, Index):
                 indexes.add(x)
+            elif isinstance(x, dict):
+                indexes.add(load_index(x))
             else:
                 indexes.add(cast("tuple[str, ...]", tuple(x)))
         return indexes
@@ -439,10 +442,10 @@ class Migrate:
                     cls._add_operator(cls._drop_index(model, index, True), upgrade, True)
                 # add indexes
                 for idx in new_indexes.difference(old_indexes):
-                    cls._add_operator(cls._add_index(model, idx, False), upgrade, True)
+                    cls._add_operator(cls._add_index(model, idx), upgrade, fk_m2m_index=True)
                 # remove indexes
                 for idx in old_indexes.difference(new_indexes):
-                    cls._add_operator(cls._drop_index(model, idx, False), upgrade, True)
+                    cls._add_operator(cls._drop_index(model, idx), upgrade, fk_m2m_index=True)
                 old_data_fields = list(
                     filter(
                         lambda x: x.get("db_field_types") is not None,
@@ -691,6 +694,16 @@ class Migrate:
         cls, model: type[Model], fields_name: Union[Iterable[str], Index], unique=False
     ) -> str:
         if isinstance(fields_name, Index):
+            if cls.dialect == "mysql":
+                # schema_generator of MySQL return a empty index sql
+                if hasattr(fields_name, "field_names"):
+                    # tortoise>=0.24
+                    fields = fields_name.field_names
+                else:
+                    # TODO: remove else when drop support for tortoise<0.24
+                    if not (fields := fields_name.fields):
+                        fields = [getattr(i, "get_sql")() for i in fields_name.expressions]
+                return cls.ddl.drop_index(model, fields, unique, name=fields_name.name)
             return cls.ddl.drop_index_by_name(
                 model, fields_name.index_name(cls.ddl.schema_generator, model)
             )
@@ -702,7 +715,29 @@ class Migrate:
         cls, model: type[Model], fields_name: Union[Iterable[str], Index], unique=False
     ) -> str:
         if isinstance(fields_name, Index):
-            return fields_name.get_sql(cls.ddl.schema_generator, model, False)
+            if cls.dialect == "mysql":
+                # schema_generator of MySQL return a empty index sql
+                if hasattr(fields_name, "field_names"):
+                    # tortoise>=0.24
+                    fields = fields_name.field_names
+                else:
+                    # TODO: remove else when drop support for tortoise<0.24
+                    if not (fields := fields_name.fields):
+                        fields = [getattr(i, "get_sql")() for i in fields_name.expressions]
+                return cls.ddl.add_index(
+                    model,
+                    fields,
+                    name=fields_name.name,
+                    index_type=fields_name.INDEX_TYPE,
+                    extra=fields_name.extra,
+                )
+            sql = fields_name.get_sql(cls.ddl.schema_generator, model, safe=True)
+            if tortoise.__version__ < "0.24":
+                sql = sql.replace("  ", " ")
+                if cls.dialect == "postgres" and (exists := "IF NOT EXISTS ") not in sql:
+                    idx = " INDEX "
+                    sql = sql.replace(idx, idx + exists)
+            return sql
         field_names = cls._resolve_fk_fields_name(model, fields_name)
         return cls.ddl.add_index(model, field_names, unique)